+++ src/main/java/com/akai/fire/sequence/EncoderMode.java
@@ -12,7 +12,7 @@ public enum EncoderMode {
             new EncoderAccess[]{NoteStepAccess.REPEATCURVE, NoteStepAccess.REPEAT_VEL_CRV, //
                     NoteStepAccess.REPEAT_VEL_END, NoteStepAccess.OCCURENCE}),
     USER_1(BiColorLightState.MODE_USER1, "1: Level\n2: Pan\n3: Fx1\n4: Fx2", new EncoderAccess[]{}),
-	USER_2(BiColorLightState.MODE_USER2, "1: Select / Tune\n2: Attack / Decay \n3: Decay / Param 1\n4: Sustain / Param 2", new EncoderAccess[]{});
+	USER_2(BiColorLightState.MODE_USER2, "1: Attack / Tune\n2: Decay / Decay \n3: Sustain / Param 1\n4: Release / Param 2", new EncoderAccess[]{});
 
     private final BiColorLightState state;
     private final String info;