+++ CHANGES.md
@@ -8,4 +8,11 @@
 - Metronome is mapped to metronome button
 - Pin track instead of pin clip
 
-- User 2 - first 4 macros from the first device in chain
+- **Encoder Modes** New mode: User 2 - first 4 macros from the first device in chain
+- **Encoder Modes** Shift + Select Mode - enable alternative mode
+
+- **Solo mode** Layer is active until mode changed 
+- **Solo mode** Keep clip colors
+
+- **Mute mode** Layer is active until mode changed 
+- **Mute mode** Keep clip colors